refactor(controllers): extract generic ProgressTracker base class
EPSSProgressTracker and NVDProgressTracker were near-identical (~85 lines
each). Extract shared singleton logic into a parameterized base class,
reducing each subclass to a few lines.
